Store setting in hours
