+++ ui/pages/settings/[tab].tsx
@@ -1,14 +1,11 @@
 import Head from "next/head";
 import { useState, useEffect } from "react";
 import { useApi } from "../../hooks/useApi";
-import { useForm } from "react-hook-form";
-import { Form, Card, Tabs, Tab } from "react-bootstrap";
-import Moment from "react-moment";
+import { Tabs, Tab } from "react-bootstrap";
 import { capitalize } from "../../components/tabs";
 import { useRouter } from "next/router";
 import { Models, Actions } from "../../utils/apiData";
-import LoadingButton from "../../components/loadingButton";
-import LockedBadge from "../../components/badges/lockedBadge";
+import SettingCard from "../../components/settings/settingCard";
 
 import ImportAndUpdateAllProfiles from "../../components/settings/importAndUpdate";
 import IdentifyingProperty from "../../components/settings/identifyingProperty";
@@ -123,99 +120,6 @@ Page.getInitialProps = async (ctx) => {
   const { settings }: Actions.SettingsList = await execApi("get", `/settings`);
   return { settings, tab };
 };
-
-function SettingCard({
-  setting,
-  updateSetting,
-  loading,
-}: {
-  setting: Models.SettingType;
-  updateSetting: Function;
-  loading: boolean;
-}) {
-  const { handleSubmit, register } = useForm();
-
-  async function onSubmit(data) {
-    setting.value = data.value;
-    updateSetting(setting);
-  }
-
-  return (
-    <>
-      <Card border={setting.variant}>
-        <Card.Body>
-          <Card.Title>{setting.title}</Card.Title>
-          {setting.type !== "boolean" ? (
-            <Card.Subtitle className="mb-2 text-muted">
-              {setting.description}
-            </Card.Subtitle>
-          ) : null}
-          <LockedBadge object={setting} />
-
-          <Form onSubmit={handleSubmit(onSubmit)} autoComplete="off">
-            <fieldset disabled={setting.locked !== null}>
-              <Form.Group>
-                {setting.type === "string" ? (
-                  <Form.Control
-                    style={{ width: "100%" }}
-                    required
-                    type="text"
-                    name="value"
-                    disabled={loading}
-                    ref={register}
-                    defaultValue={setting.value}
-                  />
-                ) : null}
-                {setting.type === "number" ? (
-                  <Form.Control
-                    style={{ width: "100%" }}
-                    required
-                    type="number"
-                    name="value"
-                    disabled={loading}
-                    ref={register}
-                    defaultValue={setting.value}
-                  />
-                ) : null}
-                {setting.type === "boolean" ? (
-                  <Form.Check
-                    type="checkbox"
-                    name="value"
-                    disabled={loading}
-                    label={setting.description}
-                    ref={register}
-                    defaultChecked={setting.value === "true"}
-                  />
-                ) : null}
-              </Form.Group>
-
-              <LoadingButton
-                style={{ marginTop: 5 }}
-                disabled={loading}
-                size="sm"
-                type="submit"
-                variant="outline-secondary"
-              >
-                Update
-              </LoadingButton>
-            </fieldset>
-          </Form>
-
-          <br />
-
-          <Card.Text>
-            <small>Default: {setting.defaultValue}</small>
-            <br />
-            <small>
-              Last Updated: <Moment fromNow>{setting.updatedAt}</Moment>
-            </small>
-          </Card.Text>
-        </Card.Body>
-      </Card>
-      <br />
-    </>
-  );
-}
 
 function ActionsTab({ errorHandler, successHandler }) {
   return (
